+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.Orderz;
+import kitchenpos.domain.order.Orderz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
